backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
celery[redis]==5.3.4
redis==5.0.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
smartapi-python==1.4.2
websocket-client==1.7.0
pyotp==2.9.0
nsepython==0.9
nsefin==0.0.5
nse-workday==0.0.3
py-vollib-vectorized==0.1.1
pandas==2.1.4
numpy==1.26.3
scipy==1.11.4
python-telegram-bot==20.7
python-dotenv==1.0.0
httpx==0.26.0
pydantic==2.5.3
pydantic-settings==2.1.0
requests==2.31.0
